requirements.txt
pycaw
pystray
Pillow
psutil
comtypes
